requirements.txt
boto3==1.17.46
django==3.2.6
django-cors-headers==3.7.0
celery==5.0.5
gunicorn==20.1.0
psycopg2-binary==2.9.1
Werkzeug==1.0.1
openpyxl==3.0.9
django-storages==1.11.1
